Moved interface into description package updated retrieveApiEnumerationDescriptionImpl to try with resource Removed duplicated dependency
